requirements-enhanced.txt
Flask>=3.0.0
opencv-python-headless>=4.8.0.76
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
openpyxl>=3.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
Werkzeug>=3.0.0

# Enhanced OCR and Image Processing
easyocr>=1.7.0
scikit-image>=0.21.0
imutils>=0.5.4 